feat: put login func in fetchAPI.js
