+++ package.json
@@ -11,8 +11,8 @@
   "dependencies": {
     "js-cookie": "^3.0.5",
     "next": "14.1.0",
-    "react": "^18",
-    "react-dom": "^18"
+    "react": "^18"    "react-dom": "^18",
+    "react-hot-toast": "^2.4.1"
   },
   "devDependencies": {
     "autoprefixer": "^10.0.1",
